requirements.txt
pandas>=2.0
requests>=2.28
django>=4.2
django-cors-headers>=4.3
python-dotenv>=1.0
